fixed Motor.h & Motor.cpp, changed example code
